Print source site CSS inline instead of enqueueing external file.
The previous approach copied parts of `Jetpack_Custom_CSS::link_tag()`, which isn't DRY or future-proof, and adds unnecessary code. Calling `Jetpack_Custom_CSS::get_css()` is a cleaner approach.

Calling `set_theme_mod()` each time the Customizer is loaded is a bit unorthodox, but is an acceptable temporary compromise. It's needed because there isn't a visual control to determine which site to switch to (if any). It should be changed to a better approach in the future once #11 is decided.

`save_imported_style()` doesn't work with the new approach yet, but it didn't work with the old one either, so it was left untouched to keep this commit within a reasonable scope.

Fixes #9