Rewrite plugin.
Issue #11 required a fundamentally different approach than the one that the previous version took, which resulted in more than 90% of the code being removed. Given that, it was easier to start fresh and merge the remaining bits from the previous version into the new code, rather than the other way around.

The name was changed to reflect the fact that it also switches the theme, and will import menus, widgets, etc in the future.

A temporary hard limit of 50 sites was set because of #4.

Fixes #1
Closes #6
Closes #8
Fixes #10
Fixes #11
Fixes #12